pyproject.toml
[tool.poetry]
name = "Ekstrastats"
version = "0.1.0"
description = ""
authors = ["Jacek Skoczek <jaca1989@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.7.2"
async-timeout = "4.0.3"
certifi = "2023.7.22"
charset-normalizer = "3.2.0"
crispy-bootstrap5 = "0.7"
dj-config-url = "0.1.1"
django = "4.2.4"
django-crispy-forms = "2.0"
django-redis = "5.4.0"
gunicorn = "21.2.0"
idna = "3.4"
packaging = "23.1"
psycopg2 = "2.9.7"
pytz = "2023.3.post1"
redis = "5.0.1"
requests = "2.31.0"
sqlparse = "0.4.4"
typing-extensions = "4.7.1"
tzdata = "2023.3"
urllib3 = "2.0.4"
whitenoise = "6.5.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
django-csp = "^3.7"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
